# docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "54Link POS Agent Banking API",
    "description": "Production API for 54Link POS Agent Banking Platform. Provides endpoints for agent management, transaction processing, fraud detection, KYC verification, settlement, and more.",
    "version": "2.0.0",
    "contact": { "name": "54Link Engineering", "email": "api@54link.com" },
    "license": { "name": "Proprietary", "url": "https://54link.com/license" }
  },
  "servers": [
    { "url": "https://api.54link.com", "description": "Production" },
    { "url": "https://staging-api.54link.com", "description": "Staging" },
    { "url": "http://localhost:3000", "description": "Local Development" }
  ],
...
  "paths": {
...
    "/api/stripe/webhook": {
      "post": {
        "tags": ["System"],
        "summary": "Stripe Webhook",
        "description": "Handles Stripe payment events (checkout.session.completed, invoice.paid, etc.).",
        "requestBody": {
          "content": { "application/json": { "schema": { "type": "object" } } }
        },
        "responses": { "200": { "description": "Webhook processed" } }
      }
    }
  },
  "components": {
...
    "securitySchemes": {
      "agentSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "agent_session",
        "description": "JWT session cookie set after agent login"
      },
      "oauthSession": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "OAuth session cookie for admin/supervisor SSO"
      }
    }
  }
}